package/gdb: change the default to GDB 15.x
This requires a bit of trickery on BR2_GDB_VERSION. Indeed, when the
target variant of gdb is enabled, but not the host variant, the
version choice is not available, and a default version is
used. However, version 15.x, which is now the default, requires GCC >=
9, so we have to be careful and use GDB 15.x by default if the
toolchain has GCC >= 9, or stick to GDB 14.x otherwise.

Of course, that will not fly long term, as at some point we'll drop
gdb 14.x, but for the time being, let's default to this older gdb
release if the target compiler is too old.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>