board/raspberrypi*: remove dtparam krnbt=on
The config.txt of the Rasperry Pi's supporting Wi-Fi and Bluetooth (i.e.
zero W, zero 2 W, 3, and 4) switch the Bluetooth function to use the
mini-UART (ttyS0) and restore UART0 (ttyAMA0) over the GPIOs 14 and 15
(as per miniuart-bt-overlay.dts[1]). All of them turn "on" the parameter
krnbt to autoprobe the bluetooth driver.

According to the README.md, the dtparam krnbt is now defaulting to "on"
since linux-raspberrypi 6.1.y[2]; it was defaulting to "off" before[3].

This removes the dtparam=krnbt=on from these config.txt files since the
value set is actually the default value.

[1]: https://github.com/raspberrypi/linux/blob/rpi-6.1.y/arch/arm/boot/dts/overlays/miniuart-bt-overlay.dts
[2]: https://github.com/raspberrypi/linux/blob/rpi-6.1.y/arch/arm/boot/dts/overlays/README#L294-L296
[3]: https://github.com/raspberrypi/linux/blob/rpi-5.10.y/arch/arm/boot/dts/overlays/README#L213-L215

Signed-off-by: Gaël PORTAY <gael.portay+rtone@gmail.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>