package/gcc-bare-metal: add multiple tuple support
Add support to gcc-bare-metal to support multiple architecture tuples
with the BR2_TOOLCHAIN_BARE_METAL_BUILDROOT_ARCH string.

To do this, custom configure, build and install commands are needed, so that
each step goes through a loop of the architecture tuples in the list.

To keep consistency with autotools, all the relevant host configurations have
been copied into the HOST_GCC_BARE_METAL_CONF_OPTS while removing any
configurations that do not apply to gcc and removing redundant configs.

autotools redundant configs covered by $(HOST_CONFIGURE_OPTS):
	CFLAGS="$$(HOST_CFLAGS)"
	LDFLAGS="$$(HOST_LDFLAGS)"

autotools configs not applicable to gcc:
	--disable-gtk-doc
	--disable-gtk-doc-html
	--disable-doc
	--disable-docs
	--disable-documentation
	--disable-debug
	--with-xmlto=no
	--with-fop=no
	--disable-nls

While we're at it: the following configuration option was incorrect:
--disable-initfini_array -> --disable-initfini-array
configure converts - to _ anyway so it makes no difference, but it's
better to be consistent.

Signed-off-by: Neal Frager <neal.frager@amd.com>
[Arnout: keep --disable-libstdcxx-pch and better explanation for
--disable-initfini-array]
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>