Revert "package/sqlite: bump version to 3.49.0"
sqlite 3.49.0 has changed its buildsystem, causing a lot of breakage
along the lines.

Although a few fixes have already been applied (and reverted in the
previous commits), there are remaining pain-points that are not trivial
to fix in a timely manner :

  - the buildsystem forces an rpath to /usr/lib, and it is not obvious
    how to prevent that (excepyt with patching),

  - there is an optional dependency to zlib that is not accounted for,
    neither in the target nor in the host variants; this causes issues
    when the build host has libz-devel (or similar) already installed,
    in whicxh case sqlite detets it and links with it; as it is missing
    an rpath to $(HOST_DIR)/lib, this causesw a build failure when
    another package pulls in zlib. See:
    https://autobuild.buildroot.org/results/1a81cb7c88a41a03767a1a201ee0628468f60766/build-end.log
    https://lore.kernel.org/buildroot/20250219195112.GA2631528@pevik/T/#mb7c7ac63902ed55623a5adf5d1db4a3f60db9fef

Fixing the second issue would seem easy enough, but zlib could depend on
ccache (when that is enabled), and ccqche depends on sqlite, so that
would be a circular dependency. And there is no option in sqlite's
configure script to forcefully disable use of zlib...

A revert is the quickest solution to unbreak for the time being, as the
next release is nearing due date. Another bump can be attempted later,
that addresses all the remaining issues.

This reverts commit db85638cea3414ca0293cbbb2352e5b0b63a2579.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Scott Fan <fancp2007@gmail.com>
Cc: Romain Naour <romain.naour@gmail.com>
Cc: Petr Vorel <petr.vorel@gmail.com>
Acked-by: Romain Naour <romain.naour@smile.fr>
Signed-off-by: Julien Olivain <ju.o@free.fr>